+++ pytorch_lightning/__init__.py
@@ -4,6 +4,18 @@ import logging
 
 from pytorch_lightning.__about__ import *  # noqa: F401, F403
 
+VERBOSE = 15  # between logging.INFO and logging.DEBUG, used for logging in production use cases
+
+
+def verbose(self, message, *args, **kws):
+    if self.isEnabledFor(VERBOSE):
+        self._log(VERBOSE, message, args, **kws)
+
+
+logging.addLevelName(VERBOSE, "VERBOSE")
+logging.verbose = verbose
+logging.Logger.verbose = verbose
+
 _root_logger = logging.getLogger()
 _logger = logging.getLogger(__name__)
 _logger.setLevel(logging.INFO)